C-ify undirected graph with amendments to quicksort, vector

+++ src/include/common/quicksort.h
@@ -23,10 +23,10 @@ inline void SwapValues(double* one, double* two) {
  * @param end_index The highest index to partition
  * @return The index position of the pivot after the partition
  */
-size_t Partition(double* a, int start_index, int end_index) {
+long int Partition(double* a, long int start_index, long int end_index) {
     double pivot_value = a[end_index];
-    int low_index = start_index;
-    for (int i = start_index; i < end_index; i++) {
+    long int low_index = start_index;
+    for (long int i = start_index; i < end_index; i++) {
         if (a[i] <= pivot_value) {
             SwapValues(&a[low_index], &a[i]);
             low_index++;
@@ -43,9 +43,9 @@ size_t Partition(double* a, int start_index, int end_index) {
  * @param startIndex The lowest index to be included in the sort
  * @param endIndex The highest index to be included in the sort
  */
-void Quicksort(double* a, int start_index, int end_index) {
+void Quicksort(double* a, long int start_index, long int end_index) {
     if (start_index < end_index) {
-        size_t pivot_index = Partition(a, start_index, end_index);
+        long int pivot_index = Partition(a, start_index, end_index);
         Quicksort(a, start_index, pivot_index - 1);
         Quicksort(a, pivot_index + 1, end_index);
     }
